requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# CythonLib\setup.py: 2
Cython == 0.27.3

# MonaGen.py: 2
# gif\combineGif.py: 1
# gif\gifMaker.py: 1
# gif\gifMaker4.py: 1,4
# steps\v1.py: 4
# steps\v2.py: 3
# steps\v3.py: 2
# steps\v4.py: 2
# steps\v5.py: 2
# steps\v6.py: 2
# steps\v7.py: 2
# steps\v8.py: 2
# test\testBlack.py: 2
# test\testChop.py: 1
# test\testCrop.py: 2
# test\testCross.py: 2
# test\testDistance.py: 1
# test\testWindow.py: 2
Pillow == 4.2.1

# steps\v1.py: 1
# steps\v2.py: 2
# steps\v3.py: 1
# test\testBlack.py: 1
# test\testDistance.py: 2
# test\testWindow.py: 1
numpy == 1.13.1
